requirements.txt
vgamepad
pywin32
psutil

pytesseract
pillow